requirements.txt
webuntis==0.3.1
ics==0.7.2
caldav==1.3.9
pytz==2024.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
